requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
loguru==0.7.2
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
gspread~=5.11.1